implement for floats

+++ src/lib/common/tree.hpp
@@ -5,8 +5,8 @@ namespace algos {
         int n;
         // parent[i] = -1 if i is root
         std::vector<int> parent;
-        int weight;
+        float weight;
 
-        Tree(int n, std::vector<int> parent, int w) : n(n), parent(std::move(parent)), weight(w) {}
+        Tree(int n, std::vector<int> parent, float w) : n(n), parent(std::move(parent)), weight(w) {}
     };
 }
